requirements.txt
streamlit>=1.40.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
ragas>=0.2.0
sentence-transformers>=3.0.0
psycopg2-binary>=2.9.9
pandas>=2.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
openai>=1.30.0
torch>=2.0.0
rank_bm25>=0.2.2
